fix tsts
